git-svn-id: svn+ssh://hgwdev.cse.ucsc.edu/projects/compbio/svnroot/hausslerlab/recon/branches/2009-09-16/src/cactus@3521 a8237c23-d22b-0410-ac1a-db40f18c2caa

+++ net.h
@@ -94,7 +94,7 @@ float event_getSubTreeBranchLength(Event *event);
 /*
  * Gets the number of events in the sub tree of the event, excluding the event itself.
  */
-float event_getSubTreeEventNumber(Event *event);
+int32_t event_getSubTreeEventNumber(Event *event);
 
 /*
  * Get number of children.
